requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
ansi2html==1.8.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asn1crypto==1.5.1
astor==0.8.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports-datetime-fromisoformat==2.0.2
bce-python-sdk==0.8.87
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.7.0
branca==0.6.0
build==1.0.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
coincurve==18.0.0
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cssselect==1.2.0
cssutils==2.7.1
CTkListbox==0.9
customtkinter==5.1.2
cycler==0.11.0
Cython==0.29.36
darkdetect==0.8.0
dash==2.14.1
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.14.1
DateTime==5.5
debugpy==1.6.7
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.7
directory-tree==0.0.3.1
distlib==0.3.6
docopt==0.6.2
downloader==0.98
easyocr==1.6.2
editdistance==0.6.2
efficientnet==1.0.0
essential-generators==1.0
et-xmlfile==1.1.0
evaluate==0.4.0
exceptiongroup==1.1.1
executing==1.2.0
ez-setup==0.9
fastjsonschema==2.17.1
ffmpeg-python==0.2.0
filelock==3.9.0
filetype==1.2.0
fiona==1.9.5
fire==0.5.0
Flask==3.0.0
flask-babel==3.1.0
flatbuffers==23.5.26
folium==0.14.0
fonttools==4.40.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
future==0.18.3
gast==0.4.0
geojson==3.0.1
geopandas==0.14.1
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h11==0.14.0
h5py==3.9.0
hdbscan==0.8.30
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.7
imageio==2.31.1
img2pdf==0.4.4
imgaug==0.4.0
importlib-metadata==6.8.0
imutils==0.5.4
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.12
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.12.0
Keras-Applications==1.0.8
keras-ocr==0.8.9
kiwisolver==1.4.4
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.2
libclang==16.0.0
lmdb==1.4.1
lxml==4.9.3
marisa-trie==1.2.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
ml-dtypes==0.2.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mutagen==1.47.0
natsort==8.2.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.26.2
nvidia-cuda-runtime-cu12==12.2.128
nvidia-pyindex==1.0.9
oauthlib==3.2.2
opencv-contrib-python==4.5.4.60
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.3.1
packaging==23.2
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pdf2docx==0.5.6
pickleshare==0.7.5
pikepdf==7.2.0
pillow==10.4.0
pipreqs==0.4.13
platformdirs==2.6.2
plotly==5.18.0
premailer==3.10.0
progress==1.6
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==3.20.0
psutil==5.9.5
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodome==3.18.0
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.15.1
PyMsgBox==1.0.9
PyMuPDF==1.24.8
PyMuPDFb==1.24.8
pynput==1.7.7
pyparsing==3.1.0
pyperclip==1.9.0
pyproj==3.6.1
pyproject_hooks==1.0.0
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.1
python-json-logger==2.0.7
python-version==0.0.2
pytils==0.4.1
pytweening==1.2.0
pytz==2023.3.post1
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
rapidfuzz==3.1.1
rarfile==4.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.18.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roboflow==1.1.48
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.9.1
Send2Trash==1.8.2
shapely==2.0.2
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.12
TA-Lib @ file:///C:/Users/mtx/TA_Lib-0.4.24-cp310-cp310-win_amd64.whl#sha256=e730d89fcfc3f0832b3495e6fa0e47ab802eb6aa0c65c5fa2a9fd267e18bdef4
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.10.1
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-text==2.10.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
tools==0.1.9
torch==2.0.0+cu117
torchvision==0.15.1+cu117
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
trio==0.22.1
trio-websocket==0.10.3
typing_extensions==4.8.0
tzdata==2023.3
ultralytics==8.3.21
ultralytics-thop==2.0.9
undetected-chromedriver==3.5.5
uri-template==1.3.0
urllib3==2.1.0
utils==1.0.2
validators==0.20.0
virtualenv==20.17.1
visualdl==2.5.3
wcwidth==0.2.6
webcolors==1.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.6.0
websockets==14.1
Werkzeug==3.0.1
wget==3.2
widgetsnbextension==4.0.7
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.13.0
xxhash==3.3.0
yarg==0.1.9
yarl==1.9.2
youtube-search-python==1.6.6
yt-dlp==2025.1.26
zipp==3.17.0
zope.interface==7.0.1
